Organize project structure as described
